feat: add layout component to render header and main page separate
